refactor: update delete_completed_jobs.py
